Add support for mDL credentials (#25)
Adds a new input /setup/inputs/mdl1 to the project
All steps can be run as usual, but with rs256 replaced by mdl1

cd setup/scripts
./run_setup.sh mdl1
cd ../../creds
cargo run --bin crescent --release --features print-trace zksetup --name mdl1
cargo run --bin crescent --release --features print-trace prove --name mdl1
cargo run --bin crescent --release --features print-trace show --name mdl1
cargo run --bin crescent --release --features print-trace verify --name mdl1



* Integrate mdl code into main workflow, part 1 (show/verify not done)


* Update /creds to use mDL


* Add circom circuits for mdl


* Update README


* Try to fix timezone issue


* Fix issuer timezone


* fix broken client helper unit test

* Implement proof validation

* Added a future TODO

* Add wasmer workaround


* Updated issuer sample

* Use config values for demo users

* snake case for issuername variable

* Added config to verifier

* verifier updates

* Add delete endpoint

---------

+++ setup/scripts/run_setup.sh
@@ -24,12 +24,26 @@ OUTPUTS_DIR=${ROOT_DIR}/generated_files/$NAME
 CIRCOM_DIR=${OUTPUTS_DIR}/circom
 INPUTS_DIR=${ROOT_DIR}/inputs/$NAME
 COPY_DEST=${ROOT_DIR}/../creds/test-vectors/$NAME
-
 LOG_FILE=${OUTPUTS_DIR}/${NAME}.log
 
 if [ ! -f ${INPUTS_DIR}/config.json ]; then
     echo "${INPUTS_DIR}/config.json is not found, aborting"
     exit -1 
+fi
+
+CREDTYPE_REGEX="\"credtype\": \"([a-z]+)\""
+if [[ `cat ${INPUTS_DIR}/config.json` =~ $CREDTYPE_REGEX ]]; then
+    CREDTYPE="${BASH_REMATCH[1]}"
+    echo "Credential type read from config.json: $CREDTYPE"
+else
+    CREDTYPE="jwt"
+    echo "Credential type not found in config.json, assuming JWT"
+fi
+
+if [ $CREDTYPE == 'mdl' ]; then 
+    CIRCOM_SRC_DIR="${ROOT_DIR}/circom-mdl"
+else
+    CIRCOM_SRC_DIR="${ROOT_DIR}/circom"
 fi
 
 
@@ -39,9 +53,9 @@ mkdir $CIRCOM_DIR 2>/dev/null  || true
 
 touch ${LOG_FILE}
 
-# Create sample issuer keys and a token
+# For JWTs, we create sample issuer keys and a token
 ALG_REGEX="\"alg\": \"([A-Z0-9]+)\""
-if [ ! -f ${INPUTS_DIR}/issuer.pub ] || [ ! -f ${INPUTS_DIR}/issuer.prv ] || [ ! -f ${INPUTS_DIR}/token.jwt ]; then
+if [ ${CREDTYPE} == 'jwt' ] && ([ ! -f ${INPUTS_DIR}/issuer.pub ] || [ ! -f ${INPUTS_DIR}/issuer.prv ] || [ ! -f ${INPUTS_DIR}/token.jwt ]); then
     rm ${INPUTS_DIR}/issuer.pub ${INPUTS_DIR}/issuer.prv ${INPUTS_DIR}/token.jwt 2>/dev/null && true 
 
     if [[ `cat ${INPUTS_DIR}/config.json` =~ $ALG_REGEX ]]; then
@@ -52,8 +66,8 @@ if [ ! -f ${INPUTS_DIR}/issuer.pub ] || [ ! -f ${INPUTS_DIR}/issuer.prv ] || [ !
     python3 scripts/jwt_sign.py ${INPUTS_DIR}/claims.json ${INPUTS_DIR}/issuer.prv  ${INPUTS_DIR}/token.jwt
 fi
 
-# Check that cicomlib is present
-if [ ! -f circom/circomlib/README.md ]; then
+# Check that circomlib is present
+if [ ! -f ${CIRCOM_SRC_DIR}/circomlib/README.md ]; then
     echo "Circomlib not found.  Run 'git submodule update --init --recursive' to get it."
     exit -1 
 fi
@@ -61,21 +75,23 @@ fi
 echo "- Generating ${NAME}_main.circom..."
 
 # Generate the circom main file.  
-python3 scripts/prepare_setup.py ${INPUTS_DIR}/config.json ${CIRCOM_DIR}/main.circom
+if [ ${CREDTYPE} != 'mdl' ]; then
+    python3 scripts/prepare_setup.py ${INPUTS_DIR}/config.json ${CIRCOM_DIR}/main.circom
+fi
 
 echo "- Compiling main.circom..."
 echo -e "\n=== circom output start ===" >> ${LOG_FILE}
 
 
 # Copy the circom files we need to the instance's circom folder.
-cp -r ${ROOT_DIR}/circom/* ${CIRCOM_DIR}/
+cp -r -L ${CIRCOM_SRC_DIR}/* ${CIRCOM_DIR}/
 
 # Compile the circom circuit.  First check if the hash of the circom files has changed, only re-compile if so. To force a re-build remove circom_files.sha256
 cd $CIRCOM_DIR
 
 set +e
 cat `find . -name "*.circom"` | sha256sum | diff circom_files.sha256 -
-if [[ $? -ne 0 ]] || [[ ! -e main_c_cpp ]] ;
+if [[ $? -ne 0 ]] || [[ ! -e main_js ]] ;
 then 
     echo "Circom files hash changed; re-compiling" >> ${LOG_FILE}
 
@@ -101,6 +117,18 @@ NUM_PUBLIC_INPUTS=$(grep -m 1 "public inputs:" "$LOG_FILE" | awk '{print $3}')
 # clean up the main.sym file as follows. Each entry is of the form #s, #w, #c, name as described in https://docs.circom.io/circom-language/formats/sym/
 awk -v max="$NUM_PUBLIC_INPUTS" -F ',' '$2 != -1 && $2 <= max {split($4, parts, "."); printf "%s,%s\n", parts[2], $2}' "${CIRCOM_DIR}/main.sym" > "${CIRCOM_DIR}/io_locations.sym"
 
+if [ ${CREDTYPE} == 'mdl' ]; then 
+    # Create the prover inputs (do it here, rather than in Rust like we do for JWTs; since the CBOR/mDL parsing code is in python) TODO: in future we should re-write it in rust
+    PROVER_INPUTS_FILE=${OUTPUTS_DIR}/prover_inputs.json
+    CRED_FILE=${INPUTS_DIR}/cred.txt
+    CONFIG_FILE=${INPUTS_DIR}/config.json
+    ISSUER_KEY_FILE=${OUTPUTS_DIR}/issuer.pub
+    cd scripts/
+    python3 prepare_mdl_prover.py ${CONFIG_FILE} ${CRED_FILE}  ${PROVER_INPUTS_FILE} ${ISSUER_KEY_FILE}
+    cd ${ROOT_DIR}
+fi
+
+
 # Copy files needed for zksetup, prove, etc..
 R1CS_FILE=${OUTPUTS_DIR}/main_c.r1cs
 WIT_GEN_FILE=${OUTPUTS_DIR}/circom/main_js/main.wasm
@@ -115,8 +143,15 @@ cp ${R1CS_FILE} ${COPY_DEST}/
 cp ${WIT_GEN_FILE} ${COPY_DEST}/ 
 cp ${SYM_FILE} ${COPY_DEST}/
 cp ${CONFIG_FILE} ${COPY_DEST}/
-cp ${TOKEN_FILE} ${COPY_DEST}/
 cp ${ISSUER_KEY_FILE} ${COPY_DEST}/
+
+if [ ${CREDTYPE} == 'jwt' ]; then
+    cp ${TOKEN_FILE} ${COPY_DEST}/
+fi
+
+if [ ${CREDTYPE} == 'mdl' ]; then 
+    cp ${PROVER_INPUTS_FILE} ${COPY_DEST}/
+fi
 
 cd scripts
 echo "Done."
